# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Job Portal REST API",
    "version": "1.0.0",
    "description": "Production-grade RESTful API for Job Portal featuring role-based access control, AI-powered matching, and candidate pipelines."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "summary": "List active jobs with filters and pagination",
        "tags": ["Jobs"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "location", "in": "query", "schema": { "type": "string" } },
          { "name": "job_type", "in": "query", "schema": { "type": "string" } },
          { "name": "sort", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of jobs" }
        }
      },
      "post": {
        "summary": "Post a new job listing (Employers/Admins)",
        "tags": ["Jobs"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "company_name", "location", "description"],
                "properties": {
                  "title": { "type": "string" },
                  "company_name": { "type": "string" },
                  "location": { "type": "string" },
                  "job_type": { "type": "string" },
                  "category": { "type": "string" },
                  "description": { "type": "string" },
                  "salary_min": { "type": "number" },
                  "salary_max": { "type": "number" },
                  "skills": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Job created" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/applications": {
      "post": {
        "summary": "Apply for a job",
        "tags": ["Applications"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["job_id"],
                "properties": {
                  "job_id": { "type": "integer" },
                  "cover_letter": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Application submitted" },
          "409": { "description": "Already applied" }
        }
      }
    },
...
    "/security/check-url": {
      "post": {
        "summary": "Multi-signal URL safety & scam verification",
        "tags": ["Security"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": { "type": "string" },
                  "company_name": { "type": "string" },
                  "company_website": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Security evaluation with risk level, indicators, and recommendations" }
        }
      }
    },
    "/security/check-job/{id}": {
      "get": {
        "summary": "Audit job listing and application destination for scam patterns",
        "tags": ["Security"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Job security audit results" }
        }
      }
    },
    "/security/report-job": {
      "post": {
        "summary": "Submit user report for suspicious job or scam recruiter",
        "tags": ["Security"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["job_id", "reason"],
                "properties": {
                  "job_id": { "type": "integer" },
                  "reason": { "type": "string", "enum": ["scam", "payment_request", "fake_company", "suspicious_link", "fake_recruiter", "misleading_information", "other"] },
                  "details": { "type": "string" },
                  "reporter_email": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Report submitted for admin review" }
        }
      }
    },
...
  }
}